fix: Save search button now behaves as expected.
