+++ .gitignore
@@ -4,7 +4,7 @@ pg*
 .DS*
 *vscode
 .turbo/
-
+vite.config.ts.timestamp*
 # Created by https://www.toptal.com/developers/gitignore/api/macos,windows,linux,visualstudiocode,vim,webstorm+all,node,yarn
 # Edit at https://www.toptal.com/developers/gitignore?templates=macos,windows,linux,visualstudiocode,vim,webstorm+all,node,yarn
